feat[agent-ms]: decompose generating single choice and gap filling
To make it possible to create an exact number of quizzes for each type and to make the generation process more controllable,
I've split the generation of single-choice questions and gap-filling exercises into two different steps.
It was necessary to change the prompts for generating the exercises, the format of the good exercises database and the TaskGenerator methods as well.